requirements.txt
apache-airflow==2.7.1
pytest==7.4.2
requests-mock==1.11.0
apache-airflow-providers-postgres==5.6.1
apache-airflow-providers-amazon==8.7.1
pandas==2.1.1
requests==2.31.0
pyarrow==13.0.0
